Describe meaning of stability stat as 90% drop in retrievability
The previous description that the stability can be combined with desired
retention to calculate the intervals is very vague and gives no intuition what a
given stability means. But the fact of it being a decay constant for a
retrievability drop from 100% to 90% is in my opinion easy to understand on an
intuitive level. Then you know that it should look very similar to the interval
statistic if you choose a retention of 90% (only difference due to fuzz,
siblings etc), if your desired retention is 80% then you can expect the
intervals to be larger than stability, and if your desired retention is 95% then
the intervals will be smaller than the stability.

The meaning of "retrievability" itself is not obvious but explained in the
dedicated statistic below.

Only downside with my suggested description is that it's a bit long for a
subtitle and creates a line wrap on my screen. I assume a more in-depth
description of the FSRS parameters will eventually happen in the manual and
in-app settings tooltips, so the statistics tooltips don't need to explain
everything. But if the subtitle is too long I would rather remove the second
sentence, i.e. the original subtitle.