NF: `assert mw` to improve typing
My looking at some code, the IDE showed a lot of warning that we
accessed a value that is not defined on None. Indeed `mw` could be
None.

By adding those assert, the IDE know that `mw` is non null.
If it were not the case, a null pointer exception would be called a
few lines later anyway.

I did it only for `mw`, as it's a very special variable in anki and my
goal was not to remove all ide's null warning.